+++ tests/System.Text.Utf8String.Tests/TestHelper.cs
@@ -22,7 +22,7 @@ namespace System.Text.Utf8.Tests
             Assert.Equal(str1.Length, str2.Length);
             for (int i = 0; i < str1.Length; i++)
             {
-                Assert.Equal(str1[i], str2[i]);
+                Assert.Equal(str1.Bytes[i], str2.Bytes[i]);
             }
         }
     }
